functionality: Given a user, report the traded volume of crypto assets between two dates

+++ src/main/java/ar/edu/unq/desapp/grupoG/backendapicryptoexchange/API/contracts/TransactionIntention/CreateTransactionIntentionRequest.java
@@ -7,7 +7,7 @@ public record CreateTransactionIntentionRequest (
     Integer crypto_amount,
     Long intention_creator_id
 ){
-    public Double finalPrice() {
-        return crypto_price * crypto_amount;
+    public Long finalPrice() {
+        return (long) (crypto_price * crypto_amount);
     }
 }
